+++ Mailer/extensions/mailer-module/src/migrations/20220613141157_add_deleted_at_to_tables.php
@@ -1,0 +1,26 @@
+<?php
+declare(strict_types=1);
+
+use Phinx\Migration\AbstractMigration;
+
+final class AddDeletedAtToTables extends AbstractMigration
+{
+    public function change(): void
+    {
+        $this->table('mail_source_template')
+            ->addColumn('deleted_at', 'datetime', ['default' => null, 'null' => true, 'after' => 'updated_at'])
+            ->update();
+
+        $this->table('mail_layouts')
+            ->addColumn('deleted_at', 'datetime', ['default' => null, 'null' => true, 'after' => 'updated_at'])
+            ->update();
+
+        $this->table('mail_types')
+            ->addColumn('deleted_at', 'datetime', ['default' => null, 'null' => true, 'after' => 'updated_at'])
+            ->update();
+
+        $this->table('mail_templates')
+            ->addColumn('deleted_at', 'datetime', ['default' => null, 'null' => true, 'after' => 'updated_at'])
+            ->update();
+    }
+}